pyproject.toml
[project]
name = "ai-job-assistant"
version = "0.1.0"
description = "AI-powered job matching and cover letter generator"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.1",
    "numpy>=2.4.3",
    "openai>=2.26.0",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2",
    "python-multipart>=0.0.22",
    "tiktoken>=0.12.0",
    "uvicorn>=0.41.0",
]
